[CAN-36] style: 캘린더 body 디자인 추가
