+++ src/styles/calendar.css
@@ -3,7 +3,7 @@
 @tailwind utilities;
 
 .fc-day {
-  @apply bg-white;
+  @apply bg-gs00;
 }
 
 /* 이벤트 스타일 */
@@ -13,7 +13,7 @@
 
 /* 오늘 날짜 기본 배경색 제거 */
 .fc-day-today {
-  @apply bg-white !important;
+  @apply bg-gs00 !important;
 }
 
 .fc-event {
@@ -28,6 +28,11 @@
 /* 각 셀의 높이 초기화 */
 .fc-daygrid-day {
   min-height: unset !important;
+  border: 0.5px solid #e2e8f0 !important;
+}
+
+.fc-col-header-cell {
+  @apply border-[0.5px] border-gs200 bg-gs50 py-2 text-14M text-gs500 !important;
 }
 
 .fc-daygrid-day-top > a {
@@ -35,5 +40,14 @@
 }
 
 .fc-daygrid-day-events {
-  line-height: 1.3 !important; /* ✅ 이벤트 간의 간격을 줄임 */
+  line-height: 1.3 !important;
+}
+
+.fc-day-other {
+  @apply bg-gs20;
+}
+
+.fc-daygrid-day.selected {
+  outline: 2px solid #3b82f6 !important;
+  outline-offset: -2px;
 }
